requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.5
openai>=1.10.0
python-dotenv>=1.0.0
pypdf>=3.17.0
chromadb>=0.4.22
sentence-transformers>=2.2.2
pandas>=2.0.0
numpy>=1.24.0
